+++ apps/ui/src/i18n.ts
@@ -12,6 +12,7 @@ import enFaults from "./locales/en/faults.json";
 import enHome from "./locales/en/home.json";
 import enNotifications from "./locales/en/notifications.json";
 import enStandardDevices from "./locales/en/standardDevices.json";
+import enSynoptics from "./locales/en/synoptics.json";
 import enTransports from "./locales/en/transports.json";
 import enUsers from "./locales/en/users.json";
 import frApps from "./locales/fr/apps.json";
@@ -26,6 +27,7 @@ import frFaults from "./locales/fr/faults.json";
 import frHome from "./locales/fr/home.json";
 import frNotifications from "./locales/fr/notifications.json";
 import frStandardDevices from "./locales/fr/standardDevices.json";
+import frSynoptics from "./locales/fr/synoptics.json";
 import frTransports from "./locales/fr/transports.json";
 import frUsers from "./locales/fr/users.json";
 
@@ -45,6 +47,7 @@ i18n.use(initReactI18next).init({
       home: enHome,
       notifications: enNotifications,
       standardDevices: enStandardDevices,
+      synoptics: enSynoptics,
       profile: enProfile,
     },
     fr: {
@@ -61,6 +64,7 @@ i18n.use(initReactI18next).init({
       home: frHome,
       notifications: frNotifications,
       standardDevices: frStandardDevices,
+      synoptics: frSynoptics,
       profile: frProfile,
     },
   },